requirements.txt
fluxer.py
python-dotenv
colorama
pymongo
